requirements/base.txt
# Core Flask
Flask==3.0.0
Werkzeug==3.0.1

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9

# Validation & Serialization
marshmallow==3.20.1
Flask-Marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Security
bcrypt==4.1.2
Flask-Bcrypt==1.0.1

# CORS
Flask-CORS==4.0.0

# Environment variables
python-dotenv==1.0.0

# UUID
shortuuid==1.0.11

# Date/Time
python-dateutil==2.8.2

# Redis (for sessions)
redis==5.0.1
Flask-Session==0.5.0

# Rate limiting
Flask-Limiter==3.5.0

# Logging
python-json-logger==2.0.7
